bank-backend/requirements.txt
fastapi
supabase
python-dotenv
uvicorn
pydantic